--- pre-onboarding-7th-2-1-4/src/lib/api/cars.ts
@@ -1,16 +1,0 @@
-import { CarParams } from "@/lib/interfaces";
-import API from "./main";
-
-export const getCars = async ({ fuelType, segment }: CarParams) => {
-  try {
-    const response = await API.get("/cars", {
-      params: {
-        fuelType,
-        segment
-      }
-    });
-    return response.data.payload;
-  } catch (error) {
-    console.log(error);
-  }
-};
