+++ pre-onboarding-7th-2-1-4/src/lib/api/index.ts
@@ -1,0 +1,25 @@
+import { Segment } from "@/lib/@types";
+import axios from "axios";
+
+const API = axios.create({
+  baseURL: " https://preonboarding.platdev.net/api/"
+});
+
+export const getCars = async ({
+  queryKey
+}: {
+  queryKey: [string, Segment];
+}) => {
+  try {
+    const segment = queryKey[1] === "ALL" ? "" : queryKey[1];
+    const response = await API.get("/cars", {
+      params: {
+        fuelType: "",
+        segment
+      }
+    });
+    return response.data.payload;
+  } catch (error) {
+    console.log(error);
+  }
+};
