--- pre-onboarding-7th-2-1-4/src/lib/api/main.ts
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const API = axios.create({
-  baseURL: " https://preonboarding.platdev.net/api/"
-});
-
-export default API;
